src/server/services/image-processing/requirements.txt
celery
redis
pillow
boto3
python-dotenv
psycopg2
flask